+++ java/sage/SageConstants.java
@@ -22,5 +22,5 @@ public final class SageConstants
     // Non-instantiable
   }
 
-  public static final int BUILD_VERSION = 1037;
+  public static final int BUILD_VERSION = 1046;
 }