+++ java/sage/epg/sd/SDRipper.java
@@ -227,7 +227,7 @@ public class SDRipper extends EPGDataSource
   {
     SDSession returnValue = null;
     BufferedReader reader = null;
-
+    
     //2025-02-28 jusjoken: switch to using sage properties for user/pass so it can be shared with clients that need it
     String propUsername = null;
     String propPassword = null;
@@ -246,7 +246,7 @@ public class SDRipper extends EPGDataSource
         }
         catch (Throwable t)
         {
-          if (Sage.DBG) System.out.println("ERROR executing server API call of:" + t);
+          if (Sage.DBG) System.out.println("SDRipper:openNewSession: ERROR executing server API call of:" + t);
           t.printStackTrace();
           propUsername = null;
           propPassword = null;
@@ -270,14 +270,14 @@ public class SDRipper extends EPGDataSource
       String auth = reader.readLine();
       if (auth == null)
       {
-        if (Sage.DBG) System.out.println("SDEPG Error: sdauth file is empty.");
+        if (Sage.DBG) System.out.println("SDRipper:openNewSession: Error: sdauth file is empty.");
         throw new SDException(SDErrors.SAGETV_NO_PASSWORD);
       }
       int split = auth.indexOf(' ');
       // If the file is not formatted correctly, it's as good as not existing.
       if (split == -1)
       {
-        if (Sage.DBG) System.out.println("SDEPG Error: sdauth file is missing a space between the username and password.");
+        if (Sage.DBG) System.out.println("SDRipper:openNewSession: Error: sdauth file is missing a space between the username and password.");
         throw new SDException(SDErrors.SAGETV_NO_PASSWORD);
       }
 
@@ -307,9 +307,9 @@ public class SDRipper extends EPGDataSource
             // This will throw an exception if there are any issues connecting.
             returnValue = new SDSageSession(propUsername, propPassword);
             authenticated = true;
-            if (Sage.DBG) System.out.println("SDEPG authenticated using prop based user/pass PASSED: username:" + propUsername + " password:" + propPassword);
+            if (Sage.DBG) System.out.println("SDRipper:openNewSession: Authenticated using prop based user/pass PASSED: username:" + propUsername + " password:" + propPassword);
         } catch (Exception e) {
-            if (Sage.DBG) System.out.println("SDEPG checking for prop based user/pass FAILED: username:" + propUsername + " password:" + propPassword);
+            if (Sage.DBG) System.out.println("SDRipper:openNewSession: Checking for prop based user/pass FAILED: username:" + propUsername + " password:" + propPassword);
         }
     }
     
@@ -321,22 +321,22 @@ public class SDRipper extends EPGDataSource
             authenticated = true;
             Sage.put(PROP_USERNAME, fileUsername);
             Sage.put(PROP_PASSWORD, filePassword);
-            if (Sage.DBG) System.out.println("SDEPG authenticated using file based user/pass PASSED: username:" + fileUsername + " password:" + filePassword);
+            if (Sage.DBG) System.out.println("SDRipper:openNewSession: Authenticated using file based user/pass PASSED: username:" + fileUsername + " password:" + filePassword);
         } catch (Exception e) {
-            if (Sage.DBG) System.out.println("SDEPG checking for file based user/pass FAILED: username:" + fileUsername + " password:" + filePassword);
+            if (Sage.DBG) System.out.println("SDRipper:openNewSession: Checking for file based user/pass FAILED: username:" + fileUsername + " password:" + filePassword);
         }
     }
 
     //we have now have tried both prop and file based user/pass
     if(!authenticated){
-        if (Sage.DBG) System.out.println("SDEPG ERROR: checking for BOTH user/pass FAILED: throwing SAGETV_NO_PASSWORD");
+        if (Sage.DBG) System.out.println("SDRipper:openNewSession: ERROR: checking for BOTH user/pass FAILED: throwing SAGETV_NO_PASSWORD");
         throw new SDException(SDErrors.SAGETV_NO_PASSWORD);
     }
     
     // We have just successfully authenticated, so this needs to be cleared so that updates can
     // start immediately.
     SDRipper.retryWait = 0;
-    if (Sage.DBG) System.out.println("SDEPG Successfully got token: " + returnValue.token);
+    if (Sage.DBG) System.out.println("SDRipper:openNewSession: Successfully got token: " + returnValue.token);
     return returnValue;
   }
 
@@ -2952,6 +2952,19 @@ public class SDRipper extends EPGDataSource
     {
       if (Sage.DBG) System.out.println("SDEPG Unable to use the Schedules Direct service at this time.");
       return false;
+    }
+    
+    try {
+          //2025-11-14 add SD Health check to avoid running if SD has the account blocked
+        if(SDUtils.isSDBlocked()){
+            if (Sage.DBG) System.out.println("SDEPG Account is blocked: Unable to use the Schedules Direct service at this time.");
+            return false;
+        } } catch (IOException ex) {
+            if (Sage.DBG) System.out.println("SDEPG Account is blocked: Unable to use the Schedules Direct service at this time. " + ex.getMessage());
+            return false;
+    } catch (SDException ex) {
+          if (Sage.DBG) System.out.println("SDEPG Account is blocked: Unable to use the Schedules Direct service at this time. " + ex.getMessage());
+          return false;
     }
 
     try