+++ java/sage/epg/sd/SDSageSession.java
@@ -22,7 +22,6 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.URL;
-import java.util.Random;
 
 public class SDSageSession extends SDSession
 {